refactor: extract postcss processor behind a resolver/dispatcher
